Cache RSVPs from Meetup per event
Introduce the Cache component and use a filesystem adapter to cache the
results from the query to the Meetup API.

This will prevent duplicate API calls being made if we need to run it
multiple times in a short space of time - i.e. if we want to pick more
than one winner during the same event.

Fixes #1.